+++ packages/wrapper/web/src/EngineWrapper.tsx
@@ -1,12 +1,8 @@
-/* eslint-disable react/prop-types */
-import React, { Children, Fragment, ReactElement, ElementType, FunctionComponent } from 'react';
-import ReactDOM from 'react-dom';
-import FullscreenRenderer from './FullscreenRenderer';
-import ModalRenderer from './ModalRenderer';
+import React, { Children, Fragment, ReactElement, forwardRef } from 'react';
+import { RichContentProps, ForwardedRef } from './RichContentProps';
+import FullscreenProvider from './FullscreenProvider';
+import ModalDialogProvider from './ModalDialogProvider';
 import { merge } from 'lodash';
-import { EditorState } from 'draft-js';
-import { RichContentProps } from './RichContentProps';
-import { RichContentEditor } from 'wix-rich-content-editor';
 
 interface Props {
   rcProps?: RichContentProps;
@@ -15,109 +11,48 @@ interface Props {
   children: ReactElement;
   isEditor?: boolean;
   isMobile?: boolean;
+  forwardedRef?: ForwardedRef;
   textToolbarType?: TextToolbarType;
   textToolbarContainer?: HTMLElement;
 }
 
 interface State {
-  ModalityProvider: typeof Fragment | typeof ModalRenderer | typeof FullscreenRenderer;
-  editorState?: EditorState;
-  MobileToolbar?: ElementType;
-  TextToolbar?: ElementType;
+  ModalityProvider: typeof Fragment | typeof ModalDialogProvider | typeof FullscreenProvider;
 }
 
 class EngineWrapper extends React.Component<Props, State> {
-  editor: typeof RichContentEditor;
-
-  constructor(props) {
+  constructor(props: Props) {
     super(props);
     this.state = this.stateFromProps(props);
-    if (props.isEditor) {
-      import(
-        // eslint-disable-next-line max-len
-        /* webpackChunkName: "rce-editorStateConversion"  */ `wix-rich-content-editor/dist/lib/editorStateConversion`
-      ).then(module => this.setState({ editorState: module.createEmpty() }));
-    }
   }
 
-  stateFromProps(props) {
+  stateFromProps(props: Props) {
     const { isEditor, children } = props;
     const { closeModal, openModal, onExpand } = children.props?.helpers || {};
     if (isEditor && !closeModal && !openModal) {
-      return { ModalityProvider: ModalRenderer };
+      return { ModalityProvider: ModalDialogProvider };
     } else if (!isEditor && !onExpand) {
-      return { ModalityProvider: FullscreenRenderer };
+      return { ModalityProvider: FullscreenProvider };
     }
     return { ModalityProvider: Fragment };
   }
 
-  editorRef = { ref: editor => (this.editor = editor) };
-
-  componentDidMount() {
-    if (this.editor) {
-      const { MobileToolbar, TextToolbar } = this.editor.getToolbars();
-      this.setState({ MobileToolbar, TextToolbar });
-    }
-  }
-
-  handleChange = editorState => {
-    this.setState({ editorState });
-  };
-
   render() {
-    const {
-      rcProps,
-      children,
-      isEditor,
-      isMobile,
-      textToolbarType,
-      textToolbarContainer,
-    } = this.props;
-    const { ModalityProvider, MobileToolbar, TextToolbar } = this.state;
-    let editorRef = {};
+    const { rcProps, children, forwardedRef } = this.props;
+    const { ModalityProvider } = this.state;
 
-    const wrapperProps: { isMobile?: boolean; textToolbarType?: TextToolbarType } = {
-      isMobile,
-      textToolbarType: isMobile ? 'inline' : textToolbarType,
-    };
-
-    const mergedRCProps = merge(rcProps, wrapperProps, children.props);
-
-    if (isEditor) {
-      const { onChange } = mergedRCProps;
-      mergedRCProps.onChange = editorState => {
-        onChange?.(editorState);
-        this.handleChange(editorState);
-      };
-      editorRef = this.editorRef;
-    }
-
-    const StaticToolbar = MobileToolbar || TextToolbar;
+    const mergedRCProps = merge(rcProps, children.props);
 
     return (
-      <Fragment>
-        <StaticToolbarPortal
-          StaticToolbar={StaticToolbar}
-          textToolbarContainer={textToolbarContainer}
-        />
-        <ModalityProvider {...mergedRCProps}>
-          {Children.only(React.cloneElement(children, { ...mergedRCProps, ...editorRef }))}
-        </ModalityProvider>
-      </Fragment>
+      <ModalityProvider {...mergedRCProps} ref={forwardedRef}>
+        {Children.only(React.cloneElement(children, { ...mergedRCProps }))}
+      </ModalityProvider>
     );
   }
 }
 
-const StaticToolbarPortal: FunctionComponent<{
-  StaticToolbar?: ElementType;
-  textToolbarContainer?: HTMLElement;
-}> = ({ StaticToolbar, textToolbarContainer }) => {
-  if (!StaticToolbar) return null;
-
-  if (textToolbarContainer) {
-    return ReactDOM.createPortal(<StaticToolbar />, textToolbarContainer);
-  }
-  return <StaticToolbar />;
-};
-
-export default EngineWrapper;
+export default forwardRef((props: Props, ref: ForwardedRef) => (
+  <EngineWrapper {...props} forwardedRef={ref}>
+    {props.children}
+  </EngineWrapper>
+));