+++ packages/wrapper/web/src/RichContentProps.ts
@@ -1,11 +1,11 @@
-import { EditorState } from 'draft-js';
+import { EditorState, RawDraftContentState } from 'draft-js';
 export interface RichContentProps {
   locale?: string;
   localeResource?: object;
   placeholder?: string;
   editorKey?: string;
   onChange?(editorState: EditorState): void;
-  initialState?: InitialState;
+  initialState?: RawDraftContentState;
   theme?: object;
   config?: object;
   plugins?: PluginConfig[];
@@ -13,3 +13,10 @@ export interface RichContentProps {
   helpers?: Helpers;
   [propName: string]: any;
 }
+
+export interface EditorDataInstance {
+  getContentState: () => RawDraftContentState;
+  refresh: (editorState: EditorState) => void;
+}
+
+export type ForwardedRef = any;