+++ packages/wrapper/web/src/RichContentWrapper.spec.tsx
@@ -49,40 +49,48 @@ describe('Wrapper', () => {
 
   describe('Editor', () => {
     it('should render locale="en" if unspecified', () => {
-      const element = mount(wrapper().withEditor());
+      const element = shallow(wrapper({ isEditor: true }).withEditor()).dive();
       expect(element.props()).toHaveProperty('locale');
       expect(element.props().locale).toEqual('en');
     });
     it('should render editor child if provided', () => {
-      const element = mount(wrapper().withEditor());
+      const element = shallow(wrapper({ isEditor: true }).withEditor());
       expect(element.props()).toHaveProperty('children');
     });
     it('should render with pluginsStrategy output', () => {
-      const element = shallow(wrapper({ plugins }).withEditor());
-      const instance = element.dive().instance();
+      const element = shallow(wrapper({ isEditor: true, plugins }).withEditor());
+      const instance = element
+        .dive()
+        .dive()
+        .dive()
+        .instance();
       const renderResult = instance.render();
-      const editorProps = renderResult.props.children[1].props;
+      const editorProps = renderResult.props.children.props;
       expect(editorProps).toHaveProperty('config');
       expect(editorProps.config).toHaveProperty('wix-draft-plugin-hashtag');
     });
     it('should render with themeStrategy output', () => {
-      const element = shallow(wrapper({ theme: 'Default' }).withEditor());
-      const instance = element.dive().instance();
+      const element = shallow(wrapper({ isEditor: true, theme: 'Default' }).withEditor());
+      const instance = element
+        .dive()
+        .dive()
+        .dive()
+        .instance();
       const renderResult = instance.render();
-      const editorProps = renderResult.props.children[1].props;
+      const editorProps = renderResult.props.children.props;
       expect(editorProps).toHaveProperty('theme');
       expect(editorProps.theme).toHaveProperty('modalTheme');
     });
     it('should call updateLocale on componentDidMount', () => {
-      const element = shallow(wrapper({ locale: 'en' }).withEditor());
-      const instance = element.instance();
+      const element = shallow(wrapper({ isEditor: true, locale: 'en' }).withEditor());
+      const instance = element.dive().instance();
       const spyUpdate = spyOn(instance, 'updateLocale');
       instance.componentDidMount();
       expect(spyUpdate.calls.count()).toEqual(1);
     });
     it('should render localeStrategy in strategies', async () => {
-      const element = shallow(wrapper({ locale: 'he' }).withEditor());
-      const instance = element.instance();
+      const element = shallow(wrapper({ isEditor: true, locale: 'he' }).withEditor());
+      const instance = element.dive().instance();
       const renderResult = instance.render();
       await instance.updateLocale();
       const engineProps = renderResult.props;
@@ -93,7 +101,7 @@ describe('Wrapper', () => {
 
   describe('Viewer', () => {
     it('should render locale="en" if unspecified', () => {
-      const element = mount(wrapper().withViewer());
+      const element = shallow(wrapper().withViewer()).dive();
       expect(element.props()).toHaveProperty('locale');
       expect(element.props().locale).toEqual('en');
     });
@@ -103,17 +111,25 @@ describe('Wrapper', () => {
     });
     it('should render with pluginsStrategy output', () => {
       const element = shallow(wrapper({ plugins }).withViewer());
-      const instance = element.dive().instance();
+      const instance = element
+        .dive()
+        .dive()
+        .dive()
+        .instance();
       const renderResult = instance.render();
-      const viewerProps = renderResult.props.children[1].props;
+      const viewerProps = renderResult.props.children.props;
       expect(viewerProps).toHaveProperty('config');
       expect(viewerProps.config).toHaveProperty('wix-draft-plugin-hashtag');
     });
     it('should render with themeStrategy output', () => {
       const element = shallow(wrapper({ theme: 'Default' }).withViewer());
-      const instance = element.dive().instance();
+      const instance = element
+        .dive()
+        .dive()
+        .dive()
+        .instance();
       const renderResult = instance.render();
-      const viewerProps = renderResult.props.children[1].props;
+      const viewerProps = renderResult.props.children.props;
       expect(viewerProps).toHaveProperty('theme');
       expect(viewerProps).toHaveProperty('decorators');
       expect(viewerProps.theme).toHaveProperty('modalTheme');