+++ packages/wrapper/web/src/pluginsStrategy/pluginsStrategy.ts
@@ -1,5 +1,6 @@
 import { merge } from 'lodash';
 import { RichContentProps } from '../RichContentProps';
+import { RawDraftContentState } from 'draft-js';
 
 const getPluginProps = (
   isEditor: boolean,
@@ -37,7 +38,7 @@ function viewerStrategy(
   prev: ViewerPluginsStrategy,
   curr: ViewerPluginConfig,
   theme: object,
-  initialState?: InitialState
+  initialState?: RawDraftContentState
 ) {
   const { type, config, typeMapper, decorator, inlineStyleMapper } = curr;
   return {
@@ -74,7 +75,7 @@ export default function pluginsStrategy(
       inlineStyleMappers: [],
     };
     strategy = plugins.reduce(
-      (prev, curr) => viewerStrategy(prev, curr, theme, initialState as InitialState),
+      (prev, curr) => viewerStrategy(prev, curr, theme, initialState as RawDraftContentState),
       emptyStrategy
     );
   }